feat: #526 organize bundler extension classes
Move internal bundler classes used by the extension API to the bundler
folder so that we can remove or migrate it to the bundler project later.
