+++ packages/liferay-js-toolkit-core/src/bundler/__tests__/PluginLogger.test.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-3.0-or-later
  */
 
-import PluginLogger from '../plugin-logger';
+import PluginLogger from '../PluginLogger';
 
 describe('when working with messages', () => {
 	let log;
