+++ packages/liferay-js-toolkit-core/src/file/handler/Manifest.ts
@@ -6,7 +6,7 @@
 import fs from 'fs-extra';
 import path from 'path';
 
-import PkgDesc from '../../pkg-desc';
+import PkgDesc from '../../bundler/PkgDesc';
 import {
 	ManifestJson,
 	ManifestJsonModuleFlags,
